+++ src/main/java/mvp/deplog/domain/post/domain/repository/PostRepository.java
@@ -7,5 +7,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
+
     Page<Post> findByPart(Part part, Pageable pageable);
 }
