+++ src/main/java/mvp/deplog/domain/post/dto/response/PostListRes.java
@@ -1,4 +1,4 @@
-package mvp.deplog.domain.post.dto;
+package mvp.deplog.domain.post.dto.response;
 
 import lombok.Data;
 
@@ -6,9 +6,10 @@ import java.time.LocalDateTime;
 
 @Data
 public class PostListRes {
+
     private String title;   // 제목
-    private String thumb;   // 첫번째 이미지
-    private String preview; // 미리보기(첫번째 블록 내용)
+    private String previewImage;   // 첫번째 이미지
+    private String previewContent; // 미리보
     private LocalDateTime createdDate;  // 작성 날짜
     private String name;    // 작성자명
     private Integer viewCount;  // 조회수
